fifth commit

+++ src/main/java/com/example/hra/service/EmployeeServiceImplement.java
@@ -5,13 +5,12 @@ import com.example.hra.dto.TotalCommissionDTO;
 import com.example.hra.entity.Department;
 import com.example.hra.entity.Employee;
 import com.example.hra.entity.Job;
+import com.example.hra.exception.EmployeeNotFoundException;
 import com.example.hra.repository.DepartmentRepository;
 import com.example.hra.repository.EmployeeRepository;
 import com.example.hra.repository.JobRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import javax.persistence.EntityManager;
 import java.math.BigDecimal;
 import java.util.*;
 import java.util.stream.Collectors;
@@ -23,15 +22,7 @@ public class EmployeeServiceImplement implements EmployeeService {
         private  EmployeeRepository employeeRepository;
         private JobRepository jobRepository;
         private DepartmentRepository departmentRepository;
-
-        private EntityManager entityManager;
-
         @Autowired
-        public void setEntityManager(EntityManager entityManager) {
-            this.entityManager = entityManager;
-        }
-
-    @Autowired
         public void setJobRepository(JobRepository jobRepository) {
             this.jobRepository = jobRepository;
         }
@@ -49,12 +40,15 @@ public class EmployeeServiceImplement implements EmployeeService {
     public Employee addEmployee(Employee employee) {
         return employeeRepository.save(employee);
     }
-
     @Override
     public Employee updateEmployee(Employee employee) {
+        Employee employee1 = employeeRepository.findByEmployeeId(employee.getEmployeeId());
+        if(employee1==null)
+        {
+            throw new EmployeeNotFoundException("Employee Not Found");
+        }
         return employeeRepository.save(employee);
     }
-
     @Override
     public Employee assignJobToEmployee(BigDecimal employeeId, String jobId) {
         Employee employee = employeeRepository.getEmployeeByEmployeeId(employeeId);
@@ -63,7 +57,9 @@ public class EmployeeServiceImplement implements EmployeeService {
             employee.setJob(job);
             return employeeRepository.save(employee);
         }
-        return null;
+        else {
+            throw new EmployeeNotFoundException("Employee or Job Not Found");
+        }
     }
         @Override
         public Employee assignDepartmentAndSalesPercentage(BigDecimal employeeId, BigDecimal departmentId, BigDecimal salesPercentage) {
@@ -74,20 +70,36 @@ public class EmployeeServiceImplement implements EmployeeService {
                 employee.setCommissionPct(salesPercentage);
                 return employeeRepository.save(employee);
             }
-            return null;
+            else
+            {
+                throw new EmployeeNotFoundException("Employee or Department Not Found");
+            }
         }
-
         @Override
         public Employee findByFirstName(String firstName) {
-            return employeeRepository.findByFirstName(firstName);
+            Employee employee =  employeeRepository.findByFirstName(firstName);
+            if(employee==null)
+            {
+                throw new EmployeeNotFoundException("Employee Not Found");
+            }
+            return employee;
         }
-
         @Override
         public Employee findByEmail(String email) {
+            Employee employee = employeeRepository.findByPhoneNumber(email);
+            if(employee==null)
+            {
+                throw new EmployeeNotFoundException("Employee Not Found");
+            }
             return employeeRepository.findByEmail(email);
         }
         @Override
         public Employee findByPhoneNumber(String phoneNumber) {
+            Employee employee = employeeRepository.findByPhoneNumber(phoneNumber);
+            if(employee==null)
+            {
+                throw new EmployeeNotFoundException("Employee Not Found");
+            }
             return employeeRepository.findByPhoneNumber(phoneNumber);
         }
     @Override
@@ -102,30 +114,35 @@ public class EmployeeServiceImplement implements EmployeeService {
                 mp.put(title,maxSalary);
                 return mp;
             }
-        return null;
+            else
+            {
+             throw new EmployeeNotFoundException("Employee Not Found");
+            }
     }
-
     public List<Employee> findAllEmployeesByDepartmentId(BigDecimal departmentId) {
         Department department = departmentRepository.findByDepartmentId(departmentId);
         List<Employee> employees = department.getEmployees();
+        if(employees.isEmpty())
+        {
+            throw new EmployeeNotFoundException("Employees Not Found in Department");
+        }
         return employees;
     }
-
-
     public List<DepartmentEmployeeCountDTO> countAllEmployeesGroupByDepartment() {
         List<Object[]> results = employeeRepository.countEmployeesGroupByDepartment();
-
         return results.stream()
                 .map(result -> new DepartmentEmployeeCountDTO((String) result[0], (Long) result[1]))
                 .collect(Collectors.toList());
     }
-
-
     @Override
     public List<Employee> findAllEmployeesWithNoCommission() {
-        return employeeRepository.findByCommissionPctIsNull();
+        List<Employee> employees = employeeRepository.findByCommissionPctIsNull();
+        if(employees.isEmpty())
+        {
+            throw new EmployeeNotFoundException("Employees Not Found");
+        }
+        return employees;
     }
-
         @Override
         public Employee updateEmployeeEmailByEmployeeId(String email,BigDecimal employeeId) {
             Employee employee = employeeRepository.findByEmployeeId(employeeId);
@@ -133,7 +150,10 @@ public class EmployeeServiceImplement implements EmployeeService {
                 employee.setEmail(email);
                 return employeeRepository.save(employee);
             }
-            return null;
+            else
+            {
+                throw new EmployeeNotFoundException("Employee Not Found");
+            }
         }
         @Override
         public Employee updateEmployeePhoneNumberByEmployeeId( String phoneNumber,BigDecimal employeeId) {
@@ -142,11 +162,27 @@ public class EmployeeServiceImplement implements EmployeeService {
                 employee.setPhoneNumber(phoneNumber);
                 return employeeRepository.save(employee);
             }
-            return null;
+            else
+            {
+                throw new EmployeeNotFoundException("Employee Not Found");
+            }
         }
         @Override
         public void deleteEmployee(BigDecimal employeeId) {
-            employeeRepository.deleteById(employeeId);
+            Employee employee = employeeRepository.findByEmployeeId(employeeId);
+            if(employee==null)
+            {
+                    throw new EmployeeNotFoundException("Employee Not Found");
+            }
+            try
+            {
+                employeeRepository.deleteById(employeeId);
+            }
+            catch (RuntimeException re)
+            {
+                throw new EmployeeNotFoundException("Employee Can't be Deleted");
+            }
+
         }
 
     @Override
@@ -165,12 +201,12 @@ public class EmployeeServiceImplement implements EmployeeService {
         return employees1;
     }
 
+    @Override
     public List<Job> findAllOpenPositions() {
         int count = 0;
         List<Job> allJobs = jobRepository.findAll();
         List<Job> openPositions = new ArrayList<>();
         List<Employee> employees = employeeRepository.findAll();
-
         for (Job job : allJobs) {
             for (Employee e : employees) {
                 if (job.getJobId() == e.getJob().getJobId()) {
@@ -181,12 +217,20 @@ public class EmployeeServiceImplement implements EmployeeService {
                 openPositions.add(job);
             count = 0;
         }
-
+        if(openPositions.isEmpty())
+        {
+            throw new EmployeeNotFoundException("No OpenPositions");
+        }
         return openPositions;
     }
     @Override
     public List<Employee> findAllEmployeesGroupByDepartment(BigDecimal departmentId) {
-        return employeeRepository.findAllEmployeesGroupByDepartment(departmentId);
+        List<Employee> employees = employeeRepository.findAllEmployeesGroupByDepartment(departmentId);
+       if(employees.isEmpty())
+       {
+           throw new EmployeeNotFoundException("Employees Not Found");
+       }
+       return employees;
     }
 
     @Override
@@ -195,6 +239,10 @@ public class EmployeeServiceImplement implements EmployeeService {
         List<Job> allJobs = jobRepository.findAll();
         List<Job> openPositions = new ArrayList<>();
         Department department = departmentRepository.findByDepartmentId(departmentId);
+        if(department==null)
+        {
+            throw new EmployeeNotFoundException("Department Not Found");
+        }
         List<Employee> employees = department.getEmployees();
         for (Job job : allJobs) {
             for (Employee e : employees) {
@@ -206,6 +254,10 @@ public class EmployeeServiceImplement implements EmployeeService {
                 openPositions.add(job);
             count = 0;
         }
+        if(openPositions.isEmpty())
+        {
+            throw new EmployeeNotFoundException("No OpenPositions");
+        }
         return openPositions;
     }
     @Override
@@ -213,11 +265,15 @@ public class EmployeeServiceImplement implements EmployeeService {
         List<Object[]> results = employeeRepository.countEmployeesGroupByLocation();
         return results;
     }
-
     public List<Employee> listAllEmployeeByHireDateRange(Date fromHireDate, Date toHireDate) {
         return employeeRepository.findAllByHireDateBetween(fromHireDate, toHireDate);
     }
     public TotalCommissionDTO findTotalCommissionIssuedToEmployeeByDepartment(BigDecimal departmentId) {
+            Department department = departmentRepository.findByDepartmentId(departmentId);
+            if(department==null)
+            {
+                throw new EmployeeNotFoundException("Department Not Found");
+            }
         BigDecimal totalCommission = employeeRepository.calculateTotalCommissionByDepartment(departmentId);
         TotalCommissionDTO commissionDTO = new TotalCommissionDTO(departmentId, totalCommission);
         return commissionDTO;
@@ -225,31 +281,27 @@ public class EmployeeServiceImplement implements EmployeeService {
 
     public String assignDepartmentAndUpdateSalesPercentage(BigDecimal departmentId, BigDecimal salesPercentage) {
         Optional<Department> optionalDepartment = departmentRepository.findById(departmentId);
-
         if (optionalDepartment.isEmpty()) {
-            return "Department not found.";
+            throw new EmployeeNotFoundException("Department not found.");
         }
-
         Department department = optionalDepartment.get();
         if (!department.getDepartmentName().equalsIgnoreCase("Sales")) {
-            return "Only employees from the Sales department can have their sales percentage updated.";
+            throw new EmployeeNotFoundException("Only employees from the Sales department can have their sales percentage updated.");
         }
         List<Employee> salesEmployees = department.getEmployees();
         if (salesEmployees.isEmpty()) {
-            return "No employees found in the Sales department.";
+            throw new EmployeeNotFoundException("No employees found in the Sales department.");
         }
         if (salesPercentage.compareTo(BigDecimal.ZERO) < 0 || salesPercentage.compareTo(BigDecimal.valueOf(100)) > 0) {
-            return "Sales percentage must be between 0 and 100.";
+            throw new EmployeeNotFoundException("Sales percentage must be between 0 and 100.");
         }
         for (Employee employee : salesEmployees) {
             employee.setDepartment(department);
             employee.setCommissionPct(salesPercentage);
         }
-
         employeeRepository.saveAll(salesEmployees);
         return "Record Modified Successfully";
     }
-
     @Override
     public String assignManager(BigDecimal employeeId, BigDecimal managerId) {
         Employee employee = employeeRepository.findByEmployeeId(employeeId);
@@ -257,12 +309,11 @@ public class EmployeeServiceImplement implements EmployeeService {
         System.out.println(employee.getFirstName());
         System.out.println(manager.getFirstName());
         if (employee==null || manager==null) {
-            return "Employee or Manager not found.";
+            throw new EmployeeNotFoundException("Employee or Manager not found.");
         }
         if (employee.getFirstName().equals(manager.getFirstName())){
-            return "Cannot assign employee as their own manager.";
+            throw new EmployeeNotFoundException("Cannot assign employee as their own manager.");
         }
-
         employee.setManager(manager);
         employeeRepository.save(employee);
         return "Record Modified Successfully";
@@ -271,17 +322,14 @@ public class EmployeeServiceImplement implements EmployeeService {
     public String assignDepartment(BigDecimal employeeId, BigDecimal departmentId) {
         Employee employee = employeeRepository.findByEmployeeId(employeeId);
         Department department = departmentRepository.findByDepartmentId(departmentId);
-
         if (employee==null || department==null) {
-            return "Employee or Department not found.";
+            throw new EmployeeNotFoundException("Employee or Department not found.");
         }
-
         employee.setDepartment(department);
         employeeRepository.save(employee);
 
         return "Record Modified Successfully";
     }
-
 
     @Override
     public List<Employee> getAllEmployees() {
